Refactor UI and routing, adjust component usage and content
This commit covers a variety of changes. Primarily it removes an unused CatalogPage component and comments out an 'Ongoing' route in the app routing. Additionally, a new utility function, buildTimeString, is introduced to generate a time string for the HistoryCard component. Several css properties and the structure of URL in the HistoryCard and SingleComingAnimeCard components are updated. Contents in the homepage and catalog are also modified. A log statement is added into useVideoTracks.ts for debugging, and the "categories" page now uses the CatalogItems2 component instead of CatalogItems.