Refactored `IEpisodeData` interface into `IEpisode` and `IEpisodeData`
In this commit, the interface `IEpisodeData` from services and components has been split into `IEpisode` and a new form of `IEpisodeData`, making the code more organized and clearer to understand. Moreover, `User.ts` is updated with a function `getLastWatchedEpisodes` to fetch last watched episodes. A new component `HistoryListLanding` is also introduced.