Refactor `generateMetadata` function in anime watch page
The name of the episode's variable has been changed to 'episodeData' in order to avoid confusion, and logging has been added to track the metadata generation process. Consequently, all instances of the episode's variable have been updated to match the new naming convention. This update aids in code clarity and debugging processes.