Add conditional rendering for HistoryListLanding component
The HistoryListLanding component has been updated to only render if there is an existing history. Previously, the component would always render regardless of the presence of history which led to potential bugs and errors. With this change, the app now ensures that the component only gets invoked when there is a history dataset to display.