Add collections feature to Anime service
This update introduces a new collections feature to the Anime service. Two new interfaces, ICollection and ICollectionItem, are defined to support this feature. Also, a new getCollection method is added to the Anime service. With this implementation, a Collection component is created to display a collection of related animes for a given anime id.